agent/requirements.txt
mcp
azure-storage-blob
azure-identity
python-dotenv
chainlit
openai
aiohttp
fastmcp